requirements.txt
websockets
aiohttp 
aiosqlite
pytz
pandas
numpy
web3
py-clob-client
python-dotenv